Improvements to AI
This commit aims to increase the intelligence of the AI while preventing it from cheating.

What this commit brings:
AI now do an extra raycast check for vision to see if the enemy as a direct line of sight.
AI vision radius now uses a triangle shape instead of a sphere.
AI can now take damage and die.
AI now stops when in attack radius of the player.