Major Code Re-Write
The project was getting a little hard to maintain so I decided to re-write all the code for the game. I have not finished re-writing it all so this is what I have.

I will try to keep as much of the base game I can, the project is just under heavy maintenance.

No update to readme as everything such as legal documents are still the same.